+++ src/client/components/AccountNewButton/AccountNewButton.js
@@ -20,7 +20,6 @@ import './AccountNewButton.html';
 
 Template.AccountNewButton.onCreated( function(){
     const self = this;
-    //console.debug( this );
 
     self.AM = {
         amInstance: new ReactiveVar( null )