+++ src/common/classes/private/am-class-tabular.js
@@ -59,7 +59,6 @@ export const amClassTabular = {
             order: [[ fieldSet.indexByName( 'emails.$.address' ), 'asc' ]],
             destroy: true
         });
-        //console.debug( 'columns', fieldSet.toTabular());
         return tabular;
     }
 };