+++ src/common/js/configure.js
@@ -6,7 +6,10 @@
 
 import _ from 'lodash';
 
+import { Logger } from 'meteor/pwix:logger';
 import { ReactiveVar } from 'meteor/reactive-var';
+
+const logger = Logger.get();
 
 let _conf = {};
 AccountsManager._conf = new ReactiveVar( _conf );
@@ -31,17 +34,13 @@ AccountsManager.configure = function( o ){
             if( Object.keys( AccountsManager._defaults ).includes( it )){
                 built_conf[it] = o[it];
             } else {
-                console.warn( 'pwix:accounts-manager configure() ignore unmanaged key \''+it+'\'' );
+                logger.warn( 'configure() ignore unmanaged key \''+it+'\'' );
             }
         });
         if( Object.keys( built_conf ).length ){
             _conf = _.merge( AccountsManager._defaults, _conf, built_conf );
             AccountsManager._conf.set( _conf );
-            // be verbose if asked for
-            if( _conf.verbosity & AccountsManager.C.Verbose.CONFIGURE ){
-                //console.log( 'pwix:accounts-manager configure() with', o, 'building', AccountsList._conf );
-                console.log( 'pwix:accounts-manager configure() with', built_conf );
-            }
+            logger.verbose({ verbosity: _conf.verbosity, against: AccountsManager.C.Verbose.CONFIGURE }, 'configure() with', built_conf );
         }
     }
     // also acts as a getter