+++ src/common/js/index.js
@@ -2,13 +2,12 @@
  * pwix:accounts-manager/src/common/js/index.js
  */
 
-import { amClass } from '../classes/am-class.class.js';
+import { amClass } from '../classes/am-class.class';
 
 import './global.js';
 import './constants.js';
 import './i18n.js';
 import './configure.js';
-import './trace.js';
 //
 import './functions.js';