+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Update to new pwix:tabular naming convention
     - Improve AccountsList email address column to have the 'more' button inside
     - Restore (disappeared) lastConnection update on user login on 'users' collection
+    - Use pwix:logger universal logger
 
 ### 2.2.0